+++ src/all/batoto/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 <application>
     <activity
-        android:name=".all.batoto.BatoToUrlActivity"
+        android:name=".all.batoto.UrlActivity"
         android:excludeFromRecents="true"
         android:exported="true"
         android:theme="@android:style/Theme.NoDisplay">
