--- src/all/batoto/CHANGELOG.md
@@ -1,201 +1,0 @@
-## 1.3.30
-
-### Refactor
-
-* Replace CryptoJS with Native Kotlin Functions
-* Remove QuickJS dependency
-
-## 1.3.29
-
-### Refactor
-
-* Cleanup pageListParse function
-* Replace Duktape with QuickJS
-
-## 1.3.28
-
-### Features
-
-* Add mirror `batocc.com`
-* Add mirror `batotwo.com`
-* Add mirror `mangatoto.net`
-* Add mirror `mangatoto.org`
-* Add mirror `mycordant.co.uk`
-* Add mirror `dto.to`
-* Add mirror `hto.to`
-* Add mirror `mto.to`
-* Add mirror `wto.to`
-* Remove mirror `mycdhands.com`
-
-## 1.3.27
-
-### Features
-
-* Change default popular sort by `Most Views Totally`
-
-## 1.3.26
-
-### Fix
-
-* Update author and artist parsing
-
-## 1.3.25
-
-### Fix
-
-* Status parsing
-* Artist name parsing
-
-## 1.3.24
-
-### Fix
-
-* Bump versions for individual extension with URL handler activities
-
-## 1.2.23
-
-### Fix
-
-* Update pageListParse logic to handle website changes
-
-## 1.2.22
-
-### Features
-
-* Add `CHANGELOG.md` & `README.md`
-
-## 1.2.21
-
-### Fix
-
-* Update lang codes
-
-## 1.2.20
-
-### Features
-
-* Rework of search
-
-## 1.2.19
-
-### Features
-
-* Support for alternative chapter list
-* Personal lists filter
-
-## 1.2.18
-
-### Features
-
-* Utils lists filter
-* Letter matching filter
-
-## 1.2.17
-
-### Features
-
-* Add mirror `mycdhands.com`
-
-## 1.2.16
-
-### Features
-
-* Mirror support
-* URL intent updates
-
-## 1.2.15
-
-### Fix
-
-* Manga description
-
-## 1.2.14
-
-### Features
-
-* Escape entities
-
-## 1.2.13
-
-### Refactor
-
-* Replace Gson with kotlinx.serialization
-
-## 1.2.12
-
-### Fix
-
-* Infinity search
-
-## 1.2.11
-
-### Fix
-
-* No search result
-
-## 1.2.10
-
-### Features
-
-* Support for URL intent
-* Updated filters
-
-## 1.2.9
-
-### Fix
-
-* Chapter parsing
-
-## 1.2.8
-
-### Features
-
-* More chapter filtering
-
-## 1.2.7
-
-### Fix
-
-* Language filtering in latest
-* Parsing of seconds
-
-## 1.2.6
-
-### Features
-
-* Scanlator support
-
-### Fix
-
-* Date parsing
-
-## 1.2.5
-
-### Features
-
-* Update supported Language list
-
-## 1.2.4
-
-### Features
-
-* Support for excluding genres
-
-## 1.2.3
-
-### Fix
-
-* Typo in some genres
-
-## 1.2.2
-
-### Features
-
-* Reworked filter option
-
-## 1.2.1
-
-### Features
-
-* Conversion from Emerald to Bato.to
-* First version
